Update README.md
'''
Description: I created a simple Caesar's cypher that can be used to encrypt a message or
decrypt a code. Encryption or decryption can be achieved with the help of a key.
In the case below the encryption key is randomly generated between negative 5 to positive 10.
This makes it pretty easy to break the message but its still cool to see how it works.
Once you encrypt a message you can decrypt it by providing the coded message and the key. 


Some points:
1.I use random.randint(-5,10) to generate the key. One of the issues is that it does pick zero sometimes
then the code has to be rerun. 
2.users can encrypt/decrypt a message with spl characters, a blank space,upper or
lower case alpahbetes and numbers.On the Ascii table thats from 32 to 126.
3. I usedthe functions ord() to convert to ascii and chr() to convert from ascii code to charater
 
'''